fix: Ensure necessary scripts are loaded into page
In Liferay "Content Page" edit mode, Vaadin portlets are not directly
added to the main page, but loaded into an iframe requesting the same
page in preview mode, and then added to the main page through javascript,
but they are detached from iframe before initialization is completed.
This way, scripts needed to register the portlet web component are not
added to the main document, so the webcomponent is not correctly rendered.
This patch add an additional script to BootstrapHandler response that
takes care to add needed script to main page and to postpone element
registration until they are loaded.

Fixes #202